+++ packages/core/src/common/client/fragments/nfts.fragment.ts
@@ -2,43 +2,65 @@ import { gql } from 'graphql-tag';
 
 export const LEGACY_NFTS_FRAGMENT = gql`
   fragment LegacyNftData on NFTv3 {
-    collection {
-      address
-    }
     attributes {
-      displayType
       traitType
-      value
-    }
-    balance {
-      scalingFactor
+      displayType
       value
     }
     description
     id
+    name
+    symbol
+    isNftSpam
+    contractType
+    spamScore
+    owner
+    collectionV3 {
+      name
+      address
+      symbol
+      collectionItemsAmount
+      collectionItemsOwnersAmount
+      floorPrices {
+        marketplaceId
+        value
+        valueUsdCents
+        paymentToken {
+          paymentTokenId
+          name
+          symbol
+          address
+          decimals
+        }
+      }
+      media {
+        type
+        url
+      }
+      marketplaces {
+        collectionUrl
+        marketplaceName
+        nftUrl
+        logoUrl
+      }
+    }
     media {
-      contentType
       type
       url
     }
-    name
-    owner
-    location
-    symbol
     lastSale {
+      quantity {
+        value
+      }
       fiatPrice {
         value
-        scalingFactor
       }
       cryptoPrice {
         value
-        scalingFactor
-      }
-      quantity {
-        value
-        scalingFactor
       }
     }
-    contractType
+    balance {
+      value
+    }
   }
 `;