+++ packages/cosmos/src/signers/private-key.signer.ts
@@ -1,7 +1,7 @@
 import { Signer, SignerDecorator } from '@xdefi-tech/chains-core';
 import { DirectSecp256k1Wallet } from '@cosmjs/proto-signing';
 import { SigningStargateClient } from '@cosmjs/stargate';
-import { fromHex } from '@cosmjs/encoding';
+import { fromBase64, fromHex } from '@cosmjs/encoding';
 import { TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
 import { bech32 } from 'bech32';
 import {
@@ -11,6 +11,7 @@ import {
 import { utils, Wallet } from 'ethers';
 import { AccAddress, RawKey, LCDClient } from '@terra-money/feather.js';
 import { encode } from 'bech32-buffer';
+import { verifyADR36Amino } from '@keplr-wallet/cosmos';
 
 import { ChainMsg, CosmosChainType } from '../msg';
 import { STARGATE_CLIENT_OPTIONS } from '../utils';
@@ -129,6 +130,27 @@ export class PrivateKeySigner extends Signer.Provider {
       });
 
       msg.sign(Buffer.from(tx.toBytes()).toString('base64'));
+    }
+  }
+
+  async verifyMessage(
+    signer: string,
+    data: Uint8Array | string,
+    pubKey: Uint8Array,
+    signature: Uint8Array
+  ): Promise<boolean> {
+    try {
+      const isVerified = verifyADR36Amino(
+        bech32.decode(signer).prefix, // prefix
+        signer, // signer
+        data, // data sign message
+        pubKey, // pubKeyBuffer
+        signature // signature
+      );
+
+      return isVerified;
+    } catch (err) {
+      return false;
     }
   }
 }