+++ packages/cosmos/src/utils.ts
@@ -7,6 +7,10 @@ import { AminoTypes } from '@cosmjs/stargate/build/aminotypes';
 import { toUtf8 } from '@cosmjs/encoding';
 import { MsgExecuteContract } from 'cosmjs-types/cosmwasm/wasm/v1/tx';
 import { Registry } from '@cosmjs/proto-signing';
+import {
+  osmosisProtoRegistry,
+  osmosisAminoConverters,
+} from 'osmojs/dist/codegen/osmosis/client';
 
 /**
  * The Amino JSON representation of [MsgExecuteContract].