+++ packages/thor/src/signers/ledger.signer.spec.ts
@@ -74,6 +74,12 @@ describe('ledger.signer', () => {
     expect(signer.verifyAddress('0xDEADBEEF')).toBe(false);
   });
 
+  it('should return false when verifing an address without prefix(thor, maya)', async () => {
+    expect(
+      signer.verifyAddress('test1hccrcavupf7wnl2klud40lan00zp0q3u807g94')
+    ).toBe(false);
+  });
+
   it('should validate an address', async () => {
     expect(signer.verifyAddress(txInput.from)).toBe(true);
   });