+++ packages/binance/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @xdefi-tech/chains-binance
 
+## 2.0.10
+
+### Patch Changes
+
+- 74a48b0: Feat: update legacyNFTs franment in core lib
+- Updated dependencies [74a48b0]
+  - @xdefi-tech/chains-core@2.0.11
+
+## 2.0.9
+
+### Patch Changes
+
+- fdbbdd7: Feat: add price.dayPriceChange field for getting balances
+- Updated dependencies [fdbbdd7]
+  - @xdefi-tech/chains-graphql@1.2.12
+
 ## 2.0.8
 
 ### Patch Changes