+++ utility-packages/graphql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @xdefi-tech/chains-graphql
 
+## 1.2.12
+
+### Patch Changes
+
+- fdbbdd7: Feat: add price.dayPriceChange field for getting balances
+
 ## 1.2.11
 
 ### Patch Changes